% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flashcards}
\alias{flashcards}
\title{Flashcard data}
\format{A data.frame with 5 variables: date, student, card, count, and
  minutes.}
\usage{
flashcards
}
\description{
Each flashcard session, 5 minutes were given to correctly identifying the
addition, subtraction, multiplication, and division answer for the 100 (90
for division) containing the single decimal digits, e.g. for addition the
cards ranged from 0+0 to 9+9. The cards were haphazardly shuffled and the
number of correctly identified result is record with a maximum of 100 in
which case the total time for all 100 cards was recorded. The flash cards
for incorrect answers were moved to the back to allow the student extra
attempts if time permitted.
}
\author{
Jarad Niemi, \email{niemi@iastate.edu}
}
\keyword{datasets}

